requirements.txt
mysql-connector-python==8.2.0
ttkbootstrap==1.10.1
tkinter
datetime
